refactor validation utils folders structure
